+++ src/test/java/kitchenpos/ordertable/application/OrderTableServiceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.ordertable.application;
 
 import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.application.OrderRepository;
 import kitchenpos.order.application.OrderService;
+import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.tablegroup.domain.OrderTableRepository;
 import kitchenpos.common.ErrorCode;
 import kitchenpos.menu.domain.Menu;
